+++ sponsorship_tracking/view/contract_view.xml
@@ -6,13 +6,28 @@
 -->
 <openerp>
 	<data>
+        <record id="view_tracking_contract_form" model="ir.ui.view">
+			<field name="name">follow.contract.compassion.form</field>
+			<field name="model">recurring.contract</field>
+            <field name="inherit_id" ref="recurring_contract.view_recurring_contract_form"/>
+			<field name="arch" type="xml">
+                <div class="oe_title" position="before">
+                    <div class="oe_right oe_button_box" name="button_box">
+                        <button name="switch_contract_view" 
+                                type="object" 
+                                string="Switch view" context="{'view_id':'view_follow_contract_form_compassion'}" />
+                    </div>
+                </div>
+            </field>
+        </record>
+        
         <record id="view_follow_contract_form_compassion" model="ir.ui.view">
 			<field name="name">follow.contract.compassion.form</field>
 			<field name="model">recurring.contract</field>
 			<field name="arch" type="xml">
                 <form string="Follow contract" version="7.0" create="false" edit="false">
                     <header>
-                        <field name="sds_state" widget="statusbar" statusbar_visible="draft,start,waiting_welcome,active,sub,sub_accept" />
+                        <field name="sds_state" widget="statusbar" statusbar_visible="draft,start,waiting_welcome,active,sub,sub_accept"/>
                         <button name="contract_validated" class="oe_highlight" attrs="{'invisible':[('sds_state','!=','draft')]}" string="Validate"/>
                         <button name="%(sponsorship_compassion.action_end_sponsorship)d" class="oe_highlight" string="Terminate" states="waiting,mandate,active" type="action"/>
                         <button name="mail_sent" class="oe_highlight" string="Mail sent" attrs="{'invisible':[('sds_state','!=','start')]}"/>
@@ -33,8 +48,12 @@
                         <div class="oe_right oe_button_box" name="button_box">
 							<button name="switch_contract_view" 
                                     type="object" 
-                                    string="Switch view"  />
+                                    string="Switch view" 
+                                    context="{'view_id':'view_tracking_contract_form'}"/>
 						</div>
+                        <div class="oe_title">
+                            <h1><field name="name"/></h1>
+                        </div>
                         <group>
                             <group>
                                 <field name="partner_id"/>
@@ -44,12 +63,13 @@
                             <group>
                                 <field name="origin_id"/>
                                 <field name="channel"/>
+                                <field name="attribution_instructions"/>
                             </group>
 
                             <group>
                                 <field name="start_date"/>
                                 <field name="activation_date"/>
-                                <field name="date_welcome"/>
+                                <field name="last_sds_state_change_date"/>
                                 <field name="end_date"/>
                             </group>
                             <group>
@@ -60,6 +80,10 @@
                             </group>
                         </group>
                     </sheet>
+                    <div class="oe_chatter">
+                        <field name="message_follower_ids" widget="mail_followers"/>
+                        <field name="message_ids" widget="mail_thread"/>
+                    </div>
                 </form>
 			</field>
 		</record>
@@ -77,7 +101,7 @@
                     <field name="child_name"/>
                     
                     <field name="start_date"/>
-                    <field name="date_welcome"/>
+                    <field name="last_sds_state_change_date"/>
                     <field name="end_date"/>
                     <field name="origin_id"/>
                     <field name="state"/>
@@ -111,7 +135,7 @@
                         <filter string="Sub date" domain="[]" context="{'group_by':'date_sub'}"/>
                         <filter string="Status" domain="[]"  context="{'group_by':'state'}"/>
                         <filter string="SDS Status" domain="[]"  context="{'group_by':'sds_state'}"/>
-                        <filter string="Project Status" domain="[]"  context="{'group_by':'sds_state'}"/>
+                        <filter string="Project Status" domain="[]"  context="{'group_by':'project_state'}"/>
                     </group>
                 </search>
             </field>
