+++ sponsorship_tracking/wizard/__init__.py
@@ -9,5 +9,4 @@
 #
 ##############################################################################
 
-from . import model
-from . import wizard
+from . import install_sds_tracking
